MenuPickerStyleで再構築
